feat: aggregate some packages
